Include incidents settings also for staging tests if mixed results are available
In the current code, if results for RRID are available, only these are
used.
This causes problem if creation of RRID changes only one arch, only a
part of tests is started and the rest reuses staging results.

Cons is we can have mixed result from older incident release which adds
complexity to https://progress.opensuse.org/issues/109974 (tough
probability is pretty low)