Make test_approve not rely on specific log level selection
Unit tests should not be too specific and we should be quite limited
with looking into actual log messages. This commit relaxes the checks
using caplog to the point what should be actually necessary for each
individual test step and not more. With this one can select an arbitrary
log level when calling pytest for easier development.