Add container definition and quadlets
As a proof of concept for running qem-bot in Podman container on MicroOS
and to support testing features from poo#180809, the gitea-sync action
service is added first. This works best together with the qem-dashboard
feature from https://github.com/openSUSE/qem-dashboard/pull/2472 as both
services can be running inside one pod.

Although this is independent on #195, the service itself calls the
feature which will be implemented in said PR. Additionally, the
deployment requires dashboard URL to be configurable (and set to
localhost) which is not yet the case outside of that PR.