Pass only packages mentioned in `_patchinfo` to dashboard
Having all packages built on OBS was not wanted on the last demo. Instead,
the dashboard should only show packages mentioned in the `_patchinfo` files
changes by the PR. The packages on OBS are of course still considered when
it comes to deciding whether a PR is ready for testing.

Related ticket: https://progress.opensuse.org/issues/180809