Accept explicitly specified single YAML file regardless of extension
When specifying a single file explicitly via the `-c` parameter it makes
sense to assume the user really wants to use that file - regardless of the
extension. This makes it easier to supply a single file on GitLab as
variable of type file.

Related ticket: https://progress.opensuse.org/issues/180812