Avoid `inconsistent scminfo …` log messages
It is expected that the build info of the different projects on OBS have
different scminfo values.

This changes tracks these scminfo values separately. The scminfo of the
codestream project is assigned as `scminfo` and the scminfo of sub projects
assigned as e.g. `scminfo_SLES`. That latter is of course not supported by
the dashboard and thus so far not actually synced there. Not sure whether
this is wanted.

If we actually want to consider the scminfo of e.g. the `SLES` sub project
as main scminfo we can easily flip the logic and adapt the test change. I'm
just going here for *something* not really knowing what the requirements
are.