Use product-specific repositories for SLFO/Gitea incidents
This is not configurable via metadata because metadata only seems to come
into play when scheduling products. However, here we need to be aware of
whether/which products to schedule already from the beginning (when syncing
incidents from Gitea).

Hence I made this configurable via two env variables for now:

* `OBS_REPO_TYPE`: specifies the repo typo to look at, e.g. `standard` or
  `product` or an empty string to consider all repo types
* `OBS_PRODUCTS`: specifies a comma-separated list of products to consider
  and if an empty value appears on the list the code stream is considered
  as well

So with `OBS_REPO_TYPE=standard` and `OBS_PRODUCTS=` we get the current
behavior of using the code stream repository. With `OBS_REPO_TYPE=product`
and `OBS_PRODUCTS=SLES` we would use the SLES-specific repository.

Note that the product version (e.g. `15.99` as in PR 166) can be read from
the build results of OBS (from the `scmsync` tag) which we read anyway. So
there's no need to supply a mapping from e.g. `1.1.99` to `15.99`.

The meta-data allows specifying "issues". So far this would be e.g.
`BASE_TEST_ISSUES: SLFO:1.1.99` where `1.1.99` is the codestream version.
Then the bot will only schedule products for channels matching this
codestream version. This will continue to work and schedule products
regardless of the product version. However, it is now also possible to
further constrain the "issues" with the product version, e.g.
`BASE_TEST_ISSUES: SLFO:1.1.99#15.99` would only schedule products for
channels with the product version `15.99`.

To determine the available architectures of product repositories the bot
now reads the `_multibuild` info from OBS. This seems to be required
because the build info returns results like this:

```
<result project="SUSE:SLFO:1.1.99:PullRequest:166:SLES" repository="product" arch="ppc64le" code="published" state="published">
```

However, there is no actual product repository for this repository present.
The absence of `ppc64le` and `local` in the `_multibuild` XML explains why
these product repositories are missing and therefore we need to check this
file for the actually relevant archs. Otherwise the bot would run later
into errors, e.g. when computing the repo hash of these non-existent repos.

Related ticket: https://progress.opensuse.org/issues/180812