Compute repo hash only for archs specified in meta-data
The product repo `SLES-SAP-16.0` seems to be only available for x86_64 at
this point causing the processing of incidents to fail as the repo hash
for e.g. aarch64 cannot be computed.

With this change the repo hash will only be computed for the archs
specified in the meta-data file avoiding this problem.

Related ticket: https://progress.opensuse.org/issues/188328