Add command to compute repo diff to find what has changed in an increment
This is supposed to be used in `increment-approve` but for the sake of
getting started I created a separate command. For the sake of local testing
the repo data can be dumped locally so it can be used later via
`--fake-data`, e.g.:

```
./qem-bot.py -d --fake-data -t none repo-diff --dump-data
```

This currently returns a diff of 24 packages but takes 30 seconds to run on
my machine without downloads (mainly due to the XML parsing).

Related ticket: https://progress.opensuse.org/issues/190251