Add AMQP handler for request_review
Extend the amqp module with a separate queue for the gitea PRs. Following the
existing design which provides the logic and the flexibility to orchestrate
the event handling according their purpose. The existing queue is listening
for suse.openqa events and nothing has changed in this flow, except some small
tweaks, which most significant, to prevent normal execution when `dry` is
enabled.

All the events from `suse.src.pull_request_review_request.review_requested`
should activate `handle_inc_review_request` which its rensposibility is to
schedule jobs. This is true if the data contains `qam-openqa-review` in the reviewers.

Here there are a few option, depending on the way the updates are being updated.
- `do_incident_schedule` will check the dashboard and schedule everything.
- `IncrementApprover._schedule_openqa_jobs` but it is an instance method
- Likely the best option is to call `post_job` from openQAInterface

The command line is append with `--queue` in order to be able to run the amqp
command against a particular queue if this is wanted. However by default it
initializes both defined queues.

issue: https://progress.opensuse.org/issues/189942
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>