Use generator in package filtering of _extra_builds_for_package
Extract the logic in its own function to simplify the
`_extra_builds_for_package` and create a generator for the
`additional_builds`.

With this refactoring, the function keeps a clean resposibility and provides
flexible function in the open/close principles. `_filter_packages_by_name` as
such provides a generator which simply returns the additional_builds to
process.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>