Prevent error+exception about already gracefully handled skip
When we encounter "Unable to read build results of project" this is
already handled gracefully and not considered an error that we need to
fix our code for. So this commit changes the "log.exception" to
"log.info" with a message addition about the consequence by mentioning
"skipping".

Related progress issue: https://progress.opensuse.org/issues/192451