feat: Reduce overly verbose "ignoring job" log message
A run of "sync incidents" is filled with thousand of lines of "ignoring
job" log.info calls which is too verbose and makes other relevant
information less accessible. This commit reduces the verbosity to a
debug message. The according test stays unaltered as the existance of
the message is still guaranteed.