Add multiple more checks based on https://github.com/l-mb/python-refactoring-skills
This adds targets for additional code quality checks that seem to be not
covered by ruff, i.e. maintainability check and duplication checks.
complexity is already covered by C901.

"radon" only outputs violations. The Python package "xenon" would call
radon and fail if violations found but the package is not available in
openSUSE so I was implementing a very simple check with inverted grep
for "any output".